aseguradora-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
bcrypt